raftstore, storage: move FlowStatistics into raftstore/worker/pd
